Use GX_DrawDone instead of GX_WaitDrawDone
GX_DrawDone both sends the draw-done command (0x45) and waits for the interrupt, while GX_WaitDrawDone only waits for the interrupt.  Most games send the draw-done command in the fifo stream, but some (such as the wrong-title fifolog for Star Fox Adventures) do not.  This resulted in a hang on Dolphin (though, not on real hardware, which is unexpected...

This issue was introduced in 2ad14f1949613066383618d807a6a8e9c706a95f and duplicated in 694dd4402d195cf1691e63fe6810f3e2c4f019f9.